fix functions
